FIX: support for response files when using cmake as a frontend for linking.
The response file is now preserved and sent to the actual linker, instead of expanding its content (and running into the limit of command line size that required the use of the response file in the first place).
Note that the first 3 arguments must be in the command line and not in the response file (otherwise cmake command line parser interpret it as a cmakelist file, and the file would need to be regenerated to remove the arguments not forwarded to the linker).