ADD: support for creating Visual Studio solutions.
Note that "Fbuild solution" must be called manually to generate the actual solution after CMake generation is done.

I copied code to get the current platform from VS generators, it might not be the best way (but for the generated projects it is only informative so it is not very important).